Bug: Fix de-sync of the sample pipeline with the gen_lvl pipeline
The sample pipeline uses the output Data object of the generator level
pipeline, to work out efficiencies. The gen lvl pipeline is fed into the
sample pipeline. Previously this was done through a config file.

In scripts, when the params of the sample pipeline are dynamically
changed, the gen lvl needs to be kept in sync. Now the gen lvl Pipeline
object can also be fed into the sample config.

Issue with hashing came up. The Pipeline object cannot be pickled, so I
did some try/except acrobatics in hash.py as a fix. Probably a more solid
fix is to check if the object (or any object inside the Sequence)
already contains a hashed value in a more general way, then replace the
unpickleable object with this hash value.