handle compound Poisson fluctuations correctly
When maps are fluctuated with the "poisson" method, they take into account
the stored error. If the error is different from Poisson errors (sqrt(N)),
a scaled Poisson process will be used to fluctuate in such a way that
the standard deviation matches the stored standard deviation.

The standard deviation is now always conserved when fluctuating with
the 'poisson' method.