fix mceq_barr with use_honda_nominal_flux
These changes are based on the oscNext2 branch in Tom's fork. They mainly are introduction of the use_honda_nominal_flux variable and most of its ocurrences in order to fix the following:

Previously, when using the honda flux, it was overwritten with the mceq flux. The fix is to differenciate between what flux is used, as is done here:
  https://github.com/ts4051/pisa/blob/oscNext2/pisa/stages/flux/mceq_barr.py#L274
  
Also, using honda flux makes it necessary to differenciate between nu and nubar, as is done here:
  https://github.com/ts4051/pisa/blob/oscNext2/pisa/stages/flux/mceq_barr.py#L412-L416
or here, for mceq_barr_red in PISA master:
  https://github.com/icecube/pisa/edit/master/pisa/stages/flux/mceq_barr_red.py#L275

In Tom's fork, an error is raised when using honda nominal flux together with the events calc mode:
  https://github.com/ts4051/pisa/blob/oscNext2/pisa/stages/flux/mceq_barr.py#L188
I was told this error is not an issue anymore and therefore didn't include the assertion in this PR, but if someone knows more about this please comment on this PR.