fix resetting parameter values for 2. octant fit
In octant fitting, if parameter values are not reset free between the single fits, the intended behavior is that after fitting the first octant, the parameter values (except for theta_23) are reset to where the first fit started. Due to the missing deepcopy, instead, they were set to the end point of the first fit instead.